Timetable Management System v1.0.
This is a mini project done by me as team lead for our semester 3 course 'Database Management Systems and Applications Lab'. It uses a Visual Basic 6.0 UI for the front-end,which is connected to an Oracle SQL database for its back-end operations. This document contains the code and screenshots of the project.